pyproject.toml
[project]
name = "holiday-media"
version = "0.1.0"
description = "Community-composable media — a 12-episode vertical micro-series parodying Hallmark holiday rom-coms through AI companionship"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "John Swords"},
]
license = {text = "MIT"}

dependencies = [
    "PyYAML>=6.0.2",
    "blake3>=0.4.1",
    "pysubs2>=1.6.1",
    "jsonschema>=4.23.0",
    "openai>=1.0.0",
    "requests>=2.31",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "types-PyYAML>=6.0.12.20240917",
    "pre-commit>=3.7.0",
]

[project.scripts]
new-cut = "scripts.new_cut:main"
compile-cut = "scripts.compile_cut:main"
apply-overlays = "scripts.apply_overlays:main"
pack-release = "scripts.pack_release:main"
select-winners = "scripts.select_winners:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "tests"]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "types-PyYAML>=6.0.12.20240917",
    "pre-commit>=3.7.0",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
target-version = "py311"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long - handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
namespace_packages = true
explicit_package_bases = true
disable_error_code = ["no-any-return", "call-overload", "call-arg", "arg-type", "attr-defined"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

[tool.coverage.run]
source = ["scripts"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
